requirements.txt
streamlit==1.44.1
pandas==2.2.3
pytest==8.3.5
fastapi==0.115.12
uvicorn==0.34.0
mysql-connector-python==9.2.0
requests==2.32.3
pydantic==2.10.3
